Move device_auth_key to separate file
